# openapi.json
{"openapi": "3.1.0", "info": {"title": "Reader -> ePub", "description": "API for converting reader to epub", "version": "1.0.0"}, "paths": {"/jobs": {"post": {"tags": ["Jobs"], "summary": "Create Job Endpoint", "operationId": "create_job_endpoint_jobs_post", "security": [{"OAuth2PasswordBearer": []}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/JobCreate"}}}}, "responses": {"202": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/JobResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "get": {"tags": ["Jobs"], "summary": "Get Jobs", "operationId": "get_jobs_jobs_get", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "page", "in": "query", "required": false, "schema": {"type": "integer", "minimum": 1, "description": "Page number", "default": 1, "title": "Page"}, "description": "Page number"}, {"name": "per_page", "in": "query", "required": false, "schema": {"type": "integer", "maximum": 100, "minimum": 1, "description": "Items per page", "default": 20, "title": "Per Page"}, "description": "Items per page"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/JobListResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/jobs/{job_id}": {"get": {"tags": ["Jobs"], "summary": "Get Job", "operationId": "get_job_jobs__job_id__get", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "job_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Job Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/JobResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/jobs/{job_id}/download": {"get": {"tags": ["Jobs"], "summary": "Download Epub", "operationId": "download_epub_jobs__job_id__download_get", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "job_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Job Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/auth/google": {"get": {"tags": ["Auth"], "summary": "Google Login", "operationId": "google_login_auth_google_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/auth/google/authorize": {"get": {"tags": ["Auth"], "summary": "Google Authorize", "operationId": "google_authorize_auth_google_authorize_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/auth/google/callback": {"get": {"tags": ["Auth"], "summary": "Google Callback", "operationId": "google_callback_auth_google_callback_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/auth/register": {"post": {"tags": ["Auth"], "summary": "Register", "operationId": "register_auth_register_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Register"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/auth/login": {"post": {"tags": ["Auth"], "summary": "Login", "operationId": "login_auth_login_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Login"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/auth/verify": {"post": {"tags": ["Auth"], "summary": "Verify Email", "operationId": "verify_email_auth_verify_post", "parameters": [{"name": "token", "in": "query", "required": true, "schema": {"type": "string", "title": "Token"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/auth/resend-verification": {"post": {"tags": ["Auth"], "summary": "Resend Verification", "operationId": "resend_verification_auth_resend_verification_post", "parameters": [{"name": "email", "in": "query", "required": true, "schema": {"type": "string", "title": "Email"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/auth/forgot-password": {"post": {"tags": ["Auth"], "summary": "Forgot Password", "operationId": "forgot_password_auth_forgot_password_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ForgotPassword"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/auth/reset-password": {"post": {"tags": ["Auth"], "summary": "Reset Password", "operationId": "reset_password_auth_reset_password_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ResetPassword"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/auth/dropbox/authorize": {"get": {"tags": ["Auth"], "summary": "Dropbox Authorize", "operationId": "dropbox_authorize_auth_dropbox_authorize_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/auth/dropbox/callback": {"get": {"tags": ["Auth"], "summary": "Dropbox Callback", "operationId": "dropbox_callback_auth_dropbox_callback_get", "parameters": [{"name": "code", "in": "query", "required": true, "schema": {"type": "string", "title": "Code"}}, {"name": "state", "in": "query", "required": true, "schema": {"type": "string", "title": "State"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/auth/onedrive/authorize": {"get": {"tags": ["Auth"], "summary": "Onedrive Authorize", "operationId": "onedrive_authorize_auth_onedrive_authorize_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/auth/onedrive/callback": {"get": {"tags": ["Auth"], "summary": "Onedrive Callback", "operationId": "onedrive_callback_auth_onedrive_callback_get", "parameters": [{"name": "code", "in": "query", "required": true, "schema": {"type": "string", "title": "Code"}}, {"name": "state", "in": "query", "required": true, "schema": {"type": "string", "title": "State"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/me": {"get": {"tags": ["Me"], "summary": "Read Me", "operationId": "read_me_me_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserResponse"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}, "put": {"tags": ["Me"], "summary": "Update Me", "operationId": "update_me_me_put", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserUpdate"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/me/usage": {"get": {"tags": ["Me"], "summary": "Get Usage", "operationId": "get_usage_me_usage_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/me/dashboard": {"get": {"tags": ["Me"], "summary": "Dashboard", "operationId": "dashboard_me_dashboard_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/jobs/{job_id}/upload": {"post": {"tags": ["Upload"], "summary": "Upload Job To Cloud", "operationId": "upload_job_to_cloud_jobs__job_id__upload_post", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "job_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Job Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UploadRequest"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/intentions": {"post": {"tags": ["Intentions"], "summary": "Capture Intention", "operationId": "capture_intention_intentions_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/IntentionCreate"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/IntentionResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/": {"get": {"summary": "Healthcheck", "operationId": "healthcheck__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}}, "components": {"schemas": {"ForgotPassword": {"properties": {"email": {"type": "string", "title": "Email"}}, "type": "object", "required": ["email"], "title": "ForgotPassword"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "IntentionCreate": {"properties": {"tier_requested": {"type": "string", "title": "Tier Requested"}}, "type": "object", "required": ["tier_requested"], "title": "IntentionCreate"}, "IntentionResponse": {"properties": {"id": {"type": "string", "title": "Id"}, "user_id": {"type": "string", "title": "User Id"}, "tier_requested": {"type": "string", "title": "Tier Requested"}, "clicked_at": {"type": "string", "format": "date-time", "title": "Clicked At"}, "reward_granted": {"type": "boolean", "title": "Reward Granted"}}, "type": "object", "required": ["id", "user_id", "tier_requested", "clicked_at", "reward_granted"], "title": "IntentionResponse"}, "JobCreate": {"properties": {"url": {"type": "string", "maxLength": 2083, "minLength": 1, "format": "uri", "title": "Url"}}, "type": "object", "required": ["url"], "title": "JobCreate"}, "JobListResponse": {"properties": {"jobs": {"items": {"$ref": "#/components/schemas/JobResponse"}, "type": "array", "title": "Jobs"}, "total": {"type": "integer", "title": "Total"}, "page": {"type": "integer", "title": "Page"}, "per_page": {"type": "integer", "title": "Per Page"}, "pages": {"type": "integer", "title": "Pages"}}, "type": "object", "required": ["jobs", "total", "page", "per_page", "pages"], "title": "JobListResponse"}, "JobResponse": {"properties": {"id": {"type": "string", "title": "Id"}, "source_url": {"type": "string", "maxLength": 2083, "minLength": 1, "format": "uri", "title": "Source Url"}, "status": {"type": "string", "title": "Status"}, "progress": {"type": "integer", "title": "Progress"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}, "external_uploads": {"additionalProperties": true, "type": "object", "title": "External Uploads", "default": {}}}, "type": "object", "required": ["id", "source_url", "status", "progress", "created_at"], "title": "JobResponse"}, "Login": {"properties": {"email": {"type": "string", "title": "Email"}, "password": {"type": "string", "title": "Password"}}, "type": "object", "required": ["email", "password"], "title": "Login"}, "Register": {"properties": {"email": {"type": "string", "title": "Email"}, "password": {"type": "string", "title": "Password"}, "name": {"type": "string", "title": "Name"}, "is_company": {"type": "boolean", "title": "Is Company", "default": false}, "country": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Country"}, "vat_number": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Vat Number"}}, "type": "object", "required": ["email", "password", "name"], "title": "Register"}, "ResetPassword": {"properties": {"token": {"type": "string", "title": "Token"}, "new_password": {"type": "string", "title": "New Password"}}, "type": "object", "required": ["token", "new_password"], "title": "ResetPassword"}, "UploadRequest": {"properties": {"provider": {"type": "string", "title": "Provider"}}, "type": "object", "required": ["provider"], "title": "UploadRequest"}, "UserResponse": {"properties": {"id": {"type": "string", "title": "Id"}, "email": {"type": "string", "title": "Email"}, "name": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Name"}, "plan": {"type": "string", "title": "Plan"}, "credits": {"type": "integer", "title": "Credits"}, "is_beta_tester": {"type": "boolean", "title": "Is Beta Tester"}, "is_active": {"type": "boolean", "title": "Is Active"}, "is_company": {"type": "boolean", "title": "Is Company"}, "country": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Country"}, "vat_number": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Vat Number"}, "connected_providers": {"items": {"type": "string"}, "type": "array", "title": "Connected Providers", "default": []}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}}, "type": "object", "required": ["id", "email", "name", "plan", "credits", "is_beta_tester", "is_active", "is_company", "country", "vat_number", "created_at"], "title": "UserResponse"}, "UserUpdate": {"properties": {"is_company": {"anyOf": [{"type": "boolean"}, {"type": "null"}], "title": "Is Company"}, "country": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Country"}, "vat_number": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Vat Number"}}, "type": "object", "title": "UserUpdate"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}, "securitySchemes": {"OAuth2PasswordBearer": {"type": "oauth2", "flows": {"password": {"scopes": {}, "tokenUrl": "token"}}}}}}